update counter example
