+++ counter/src/stackr/machine.ts
@@ -1,18 +1,8 @@
-import { State, StateMachine } from "@stackr/sdk/machine";
-import { solidityPackedKeccak256 } from "ethers";
+import { StateMachine } from "@stackr/sdk/machine";
 
 import * as genesisState from "../../genesis-state.json";
+import { CounterState } from "./state";
 import { transitions } from "./transitions";
-
-export class CounterState extends State<number> {
-  constructor(state: number) {
-    super(state);
-  }
-
-  getRootHash() {
-    return solidityPackedKeccak256(["uint256"], [this.state]);
-  }
-}
 
 const machine = new StateMachine({
   id: "counter",
