+++ counter/src/stackr/transitions.ts
@@ -1,16 +1,18 @@
 import { STF, Transitions } from "@stackr/sdk/machine";
-import { CounterState } from "./machine";
+import { CounterState } from "./state";
 
 const increment: STF<CounterState> = {
-  handler: ({ state }) => {
+  handler: ({ state, emit }) => {
     state += 1;
+    emit({ name: "ValueAfterIncrement", value: state });
     return state;
   },
 };
 
 const decrement: STF<CounterState> = {
-  handler: ({ state }) => {
+  handler: ({ state, emit }) => {
     state -= 1;
+    emit({ name: "ValueAfterDecrement", value: state });
     return state;
   },
 };
